requirements.txt
discord.py>=2.3,<3.0
aiohttp>=3.8,<4.0
aiosqlite>=0.19,<1.0
discord-ext-voice-recv>=0.3.0
openai-whisper>=20231117
serpapi>=0.1.5
pillow>=10.0.0
pytesseract>=0.3.10
numpy>=1.24.0
